refactor: use chat templates
