+++ chatbot/bot/model/model.py
@@ -10,8 +10,3 @@ class Model(ABC):
     config_answer: Optional[Dict[str, Any]]
     type: Optional[str]
     system_template: str
-    qa_prompt_template: str
-    ctx_prompt_template: str
-    refined_ctx_prompt_template: str
-    refined_question_conversation_awareness_prompt_template: str
-    refined_answer_conversation_awareness_prompt_template: str
