+++ chatbot/chatbot_app.py
@@ -4,8 +4,7 @@ import time
 from pathlib import Path
 
 import streamlit as st
-from bot.client.client_settings import get_client, get_clients
-from bot.client.llm_client import LlmClient
+from bot.client.lama_cpp_client import LamaCppClient
 from bot.conversation.conversation_retrieval import ConversationRetrieval
 from bot.model.model_settings import get_model_setting, get_models
 from helpers.log import get_logger
@@ -14,20 +13,17 @@ logger = get_logger(__name__)
 
 
 @st.cache_resource(experimental_allow_widgets=True)
-def load_llm(llm_client: LlmClient, model_name: str, model_folder: Path) -> LlmClient:
+def load_llm(model_name: str, model_folder: Path) -> LamaCppClient:
     """
     Create a LLM session object that points to the model.
     """
     model_settings = get_model_setting(model_name)
-    clients = [client.value for client in model_settings.clients]
-    if llm_client not in clients:
-        llm_client = clients[0]
-    llm = get_client(llm_client, model_folder=model_folder, model_settings=model_settings)
+    llm = LamaCppClient(model_folder=model_folder, model_settings=model_settings)
     return llm
 
 
 @st.cache_resource()
-def load_conversational_retrieval(_llm: LlmClient) -> ConversationRetrieval:
+def load_conversational_retrieval(_llm: LamaCppClient) -> ConversationRetrieval:
     conversation_retrieval = ConversationRetrieval(_llm)
     return conversation_retrieval
 
@@ -78,12 +74,11 @@ def main(parameters) -> None:
     model_folder = root_folder / "models"
     Path(model_folder).parent.mkdir(parents=True, exist_ok=True)
 
-    client = parameters.client
     model = parameters.model
     max_new_tokens = parameters.max_new_tokens
 
     init_page(root_folder)
-    llm = load_llm(client, model, model_folder)
+    llm = load_llm(model, model_folder)
     conversational_retrieval = load_conversational_retrieval(_llm=llm)
     init_chat_history(conversational_retrieval)
     init_welcome_message()
@@ -119,22 +114,8 @@ def main(parameters) -> None:
 def get_args() -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="Chatbot")
 
-    client_list = get_clients()
-    default_client = client_list[0]
-
     model_list = get_models()
     default_model = model_list[0]
-
-    parser.add_argument(
-        "--client",
-        type=str,
-        choices=client_list,
-        help=f"Client to be used. Defaults to {default_client}.",
-        required=False,
-        const=default_client,
-        nargs="?",
-        default=default_client,
-    )
 
     parser.add_argument(
         "--model",
