+++ chatbot/cli/chatbot.py
@@ -3,7 +3,7 @@ import sys
 import time
 from pathlib import Path
 
-from bot.client.client_settings import get_client, get_clients
+from bot.client.lama_cpp_client import LamaCppClient
 from bot.model.model_settings import get_model_setting, get_models
 from helpers.log import get_logger
 from helpers.reader import read_input
@@ -17,22 +17,8 @@ logger = get_logger(__name__)
 def get_args() -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="Chatbot")
 
-    client_list = get_clients()
-    default_client = client_list[0]
-
     model_list = get_models()
     default_model = model_list[0]
-
-    parser.add_argument(
-        "--client",
-        type=str,
-        choices=client_list,
-        help=f"Client to be used. Defaults to {default_client}.",
-        required=False,
-        const=default_client,
-        nargs="?",
-        default=default_client,
-    )
 
     parser.add_argument(
         "--model",
@@ -87,12 +73,7 @@ def main(parameters):
     model_folder = root_folder / "models"
     Path(model_folder).parent.mkdir(parents=True, exist_ok=True)
 
-    client = parameters.client
-    clients = [client.value for client in model_settings.clients]
-    if parameters.client not in clients:
-        client = clients[0]
-
-    llm = get_client(client, model_folder=model_folder, model_settings=model_settings)
+    llm = LamaCppClient(model_folder=model_folder, model_settings=model_settings)
     loop(llm)
