+++ todo.md
@@ -2,7 +2,6 @@
 - Try Chat Templates https://medium.com/@ahmet_celebi/demystifying-chat-templates-of-llm-using-llama-cpp-and-ctransformers-f17871569cd6
 - Test Flash attention:
   - https://github.com/ggerganov/llama.cpp/pull/5021
-  - use LLAMA_CUDA since LLAMA_CUBLAS is deprecated
 - Google Search with LLM:
   - https://huggingface.co/blog/nand-tmp/google-search-with-llm
   - https://blog.nextideatech.com/how-to-use-google-search-with-langchain-openai/
